automate-gateway/middleware: rename authv2 -> authz
